♻️ (refactor): Updated all references of EditItem to Receipts
