+++ frontend/apps/slip-snapper/src/app/App.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import Home from './pages/Home';
 import ViewReports from './pages/ViewReports';
 import Profile from './pages/Profile';
-import EditItem from './pages/EditItem';
+import Receipts from './pages/Receipts';
 import EditSlip from './pages/EditSlip';
 import Register from './pages/Register';
 import Login from './pages/Login'
@@ -43,7 +43,7 @@ const App: React.FC = () => (
         <Route path="/home" component={Home} exact={true} />
         <Route path="/viewreports" component={ViewReports} exact={true} />
         <Route path="/profile" component={Profile} exact={true} />
-        <Route path="/edititem" component={EditItem} exact={true} />
+        <Route path="/receipts" component={Receipts} exact={true} />
         <Route path="/editslip" component={EditSlip} exact={true} />
         <Route path="/register" component={Register} exact={true} />
         <Route path="/login" component={Login} exact={true} />
