+++ frontend/apps/slip-snapper/src/app/Tests/Home.spec.tsx
@@ -27,7 +27,7 @@ describe('Home', () => {
     expect(Component.getByText('Recent Reports'));
     expect(Component.getByText('Reports'));
     expect(Component.getByText('Profile'));
-    expect(Component.getByText('Edit Item'));
+    expect(Component.getByText('Receipts'));
     expect(Component.getByText('Report Summary'));
   });
 });
