+++ frontend/apps/slip-snapper/src/app/pages/Receipts.tsx
